Create authRouter, link to server
