+++ api/routers/userRouter.js
@@ -1,8 +1,6 @@
 const router = require("express").Router();
-const bcrypt = require("bcryptjs");
 
 const { users, parties } = require("../../data/helpers/dbHelpers.js");
-const secret = process.env.SECRET || ".env is not working right now";
 const genericError = { err: "There was a problem processing your request" };
 
 router.get("/:id", async (req, res) => {
@@ -33,27 +31,5 @@ router.get("/:id/parties", async (req, res) => {
   }
 });
 
-router.post("/register", async (req, res) => {
-  const newUser = req.body;
-  if (newUser.username && newUser.password) {
-    try {
-      const hashword = bcrypt.hashSync(newUser.password, 12);
-      newUser.password = hashword;
-      user = await users.registerUser(newUser);
-      if (user) {
-        res.status(201).json({ message: `Welcome ${user.username}` });
-      } else {
-        res
-          .status(400)
-          .json({ err: "Couldn't register user, please check credentials" });
-      }
-    } catch (err) {
-      console.log(err);
-      res.status(400).json({ err: "Username already taken" });
-    }
-  } else {
-    res.status(400).json({ err: "Please provide username & password" });
-  }
-});
 
 module.exports = router;
